Remove Infrastructure.MessageBus module
The Infrastructure.MessageBus module was completely removed due to redesign and refactoring of the message handling mechanism. This involved deleting several classes, interfaces, extensions, and project configurations which were no longer needed. This change aims to simplify the system's structure and improve the efficiency of the message handling processes.

+++ src/Services/Identity/Command/Infrastructure.EventStore/DependencyInjection/Options/EventStoreOptions.cs
@@ -4,5 +4,6 @@ namespace Infrastructure.EventStore.DependencyInjection.Options;
 
 public record EventStoreOptions
 {
-    [Required, Range(3, 100)] public int SnapshotInterval { get; init; }
+    [Required, Range(3, 100)]
+    public ulong SnapshotInterval { get; init; }
 }
